UnitHealth and UnitManager
UnitHealth has been added, it currently only supports enemy health responsibilities. Going forward it will be less specific and support managing the player as well. 

UnitManager is responsible for registering events such as when a unit gets punched it will decide how to react. Additionally it will be responsible for triggering other events down the line. It is designed to be very extensible so we can create variants of later for specific enemies.