Song and Playlist Assets
Playlists can be made which play different songs in different game contexts such as first or last round.
Songs are assets which contain the BPM, the sound file, and the time to wait until starting the fight.
The round manager now waits for the amount of time specified until starting the round.