Experimental Changes to Beat Controller
Made the beat controller independent of frame delay which was causing it to deviate from the beats. The new system will keep track of when next beat is expected to happen. Additionally it runs independent of the initial BPM so we could dynamically change the BPM in the middle of a song if we find the need to.